Remove printf from reduced grammar
While very useful, this is unneeded complexity to get off the ground.
Due to its high utility, it should be added back in before the first
major version release.

Need to replace because the original Awk grammar actually interprets the
statement `print("")` as a print without parens, but the value is
paren-enclosed.