Convert Expression to a Trait
Mystery: For some reason, for certain parsers, the `map` combinator does
not understand how to type-inference the `Box<dyn Expression>` and make
it line up with the `ExpressionImpl` concrete type. This led me to
convert them to explicit implementations, which then allows the code to
compile.

Two blockers to finishing this:
- Making Pattern have access to the direct regex
- Allowing Regex match to know an expression is a regex