Combinator-based Operator Precedence (#27)
* Refactor: Use HOF implementation of regex parsing

Motivation: Since we are going to be introducing many operators into a
hierarchy, it is useful to genericize how the next level of operator
gets parsed as a parameter.

* Refactor: Remove unused old implementation

* Refactor: Replace addition with genericized parser

* Refactor: Replace multiplication parser with genericized version

* Refactor: Swap overall expression parsing to HOF parsers

* Refactor: Delete unused parser

* Refactor: Make field reference internally handle repetitions

The FieldReference expression can nest with itself (e.g. `$$$$1`). This
should be made explicit by its own parser.

* Refactor: Make field references a HOF

* Refactor: Use new HOF field reference parser

* Refactor: Remove unused parser

* Test: Add test for recursive field references

Test cannot be enabled yet due to lack of string-to-number coercion.