Add custom functions (#57)
* Refactor: Extract global variables into reusable context

* Refactor: Introduce StackFrame constructor

* Behavior: Add single function frame

* Introduce stack of frames

* Implement Function invocation

* Extract function to parsable value

* Move StackFrame into Function

* Move stack creation into FunctionDefinition

* Behavior: Add parsing of function invocations

* WIP: Push immutable function collection down evaluation stack

* WIP: Push Function invocation down to Expressions

* Change mutability of Context for expressions

With function invocations being a hybrid of functions and statements, we
now need to pass a mutable context into expressions since functions
themselves can mutate the Context.

* Remove now-unused functions

These were moved up to the program so that we could guarantee they would
never be borrowed mutably. The Context can be mutated, but we never need
functions to be mutated after the program has been parsed.

* Move list parsing up to overall program

* Add function parsing

* Add test for basic function invocation parsing

* Implement function invocations

* Bugfix: Function invocations didn't parse

Due to ordering of expressions being parsed, a function invocation was
being preferably parsed as a simple variable. This caused certain
programs to fail to parse correctly.

Many new tests were added in the course of debugging this bug, so
keeping those for coverage.